LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-option-element/option-selected-expected.txt


PASS not dirty
FAIL dirty, selected assert_equals: expected true but got false
FAIL dirty, not selected assert_equals: expected false but got true
